refactor: move entities to domain folder
