+++ src/domain/entities/Freight.ts
