+++ src/domain/entities/Item.ts
