+++ src/domain/entities/errors/ExpiredCoupon.ts
