+++ src/domain/entities/errors/empty-properties/InvalidEmptyID.ts
