+++ src/domain/entities/errors/empty-properties/InvalidEmptyProperty.ts
