+++ src/domain/entities/errors/empty-properties/index.ts
