+++ src/domain/entities/errors/index.ts
