+++ tests/Order.spec.ts
@@ -1,7 +1,5 @@
-import { Order } from '@/Order'
-import { Item } from '@/Item'
-import { Coupon } from '@/Coupon'
-import { ExpiredCouponError, InvalidEmptyID } from '@/domain/errors'
+import { Order, Item, Coupon } from '@/domain/entities'
+import { ExpiredCouponError, InvalidEmptyID } from '@/domain/entities/errors'
 
 const makeSut = (
   cpf: string,
